Fix a number of 64-bit errors.
- One of these errors was causing a crash. I don't actually know which
  one. I had clang find all 64-bit warnings and fixed them the way it
  suggested, and the crash went away. I then gave them a once-over to
  make sure they all looked like good fixes, and left it at that.

+++ objc-appscript/trunk/src/Appscript/utils.m
@@ -13,7 +13,7 @@ NSString *ASDescriptionForError(OSStatus err) {
 		if (errorString && ![errorString isEqual: @""])
 			return errorString;
 	}
-	return [NSString stringWithFormat: @"Mac OS error %i", err];
+	return [NSString stringWithFormat: @"Mac OS error %ld", (long)err];
 }
